Combined Redis functions and rewrote logging
