+++ src/utils/helper.py
@@ -3,10 +3,6 @@ import logging
 import sys
 import os
 
-# Set up logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
 
 def list_tasks():
     """
